MAINT: Move modules to private namespaces
