+++ doc/source/tutorial/optimize.rst
@@ -1233,7 +1233,7 @@ The problem we have can now be solved as follows:
     # visualize
     import matplotlib.pyplot as plt
     x, y = mgrid[0:1:(nx*1j), 0:1:(ny*1j)]
-    plt.pcolor(x, y, sol.x)
+    plt.pcolormesh(x, y, sol.x, shading='gouraud')
     plt.colorbar()
     plt.show()
 
@@ -1369,7 +1369,7 @@ Consider the following simple linear programming problem:
 
 We need some mathematical manipulations to convert the target problem to the form accepted by :func:`linprog`.
 
-First of all, let's consider the objective function. 
+First of all, let's consider the objective function.
 We want to maximize the objective
 function, but :func:`linprog` can only accept a minimization problem. This is easily remedied by converting the maximize
 :math:`29x_1 + 45x_2` to minimizing :math:`-29x_1 -45x_2`. Also, :math:`x_3, x_4` are not shown in the objective
