+++ scipy/signal/waveforms.py
@@ -370,7 +370,7 @@ def chirp(t, f0, t1, f1, method='linear', phi=0, vertex_zero=True):
 
     >>> def plot_spectrogram(title, w, fs):
     ...     ff, tt, Sxx = spectrogram(w, fs=fs, nperseg=256, nfft=576)
-    ...     plt.pcolormesh(tt, ff[:145], Sxx[:145], cmap='gray_r')
+    ...     plt.pcolormesh(tt, ff[:145], Sxx[:145], cmap='gray_r', shading='gouraud')
     ...     plt.title(title)
     ...     plt.xlabel('t (sec)')
     ...     plt.ylabel('Frequency (Hz)')
