+++ scipy/special/_lambertw.pxd
