+++ scipy/special/_lambertw.py
